pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "pinger-bot"
description = "Простой бот для отслеживания статуса и статистики своих MineCraft серверов."
version = "0.1.0"
license = "AGPL3"

authors = []

readme = "README.md"

repository = "https://github.com/PerchunPak/pinger-bot"

[tool.poetry.dependencies]
python = "^3.8"

# Our dependecies
nextcord = "^2.0.0-alpha"

python-decouple = "^3"
structlog = "21.5.0"
### For colors in structlog ###
### Only needed for Windows ###
colorama = { version = "*", markers="platform_system == 'Windows'" }

# Docs
sphinx                   = { version = "^4.5",  optional = true }
sphinx-autodoc-typehints = { version = "^1.18", optional = true }
sphinx-autoapi           = { version = "^1.8",  optional = true }
furo                     = { version = "*",     optional = true }
m2r2                     = { version = "^0.3",  optional = true }
tomli                    = { version = "^2.0",  optional = true }

[tool.poetry.dev-dependencies]
black = "^22"
isort = "^5.10"
pycln = "^1.3"
doc8 = "^0.11"
pre-commit = "^2.18"
mypy = "0.950"
flake8 = "^4"
flake8-docstrings = "^1"
pydocstyle = ">=4.0.0"

safety = "^1.10"

pytest = "^7.1"
pytest-cov = "^3.0"
pytest-mock = "^3.7"
pytest-testmon = "^1.3"
pytest-randomly = "^3.11"

[tool.poetry.extras]
docs = ["sphinx", "sphinx-autodoc-typehints", "sphinx-autoapi", "furo", "m2r2", "tomli"]

[tool.black]
line-length = 120
target-version = ['py310']
